renamed files to index for easier importing
